Finishes: Add model modules - Refactor the model files to reference modules for accuracy, activation, data manager, layer, loss and optimizer - rename the entry file to start.py
